refactor(biz-session): update BizSession type and related components to improve data handling and consistency across the application
