+++ apps/web/src/pages/users/components/session-detail-content.tsx
@@ -3,14 +3,33 @@ import {
   ArrowLeftIcon,
   ChevronDownIcon,
   ChevronUpIcon,
+  PieChartIcon,
 } from '@radix-ui/react-icons';
 import { Link, useNavigate } from 'react-router-dom';
 import { useQuerySessionDetailQuery } from '@usertour-ui/shared-hooks';
 import { Table, TableBody, TableCell, TableRow } from '@usertour-ui/table';
-import { BizEvent, BizEvents } from '@usertour-ui/types';
+import { BizEvent, BizEvents, ContentDataType } from '@usertour-ui/types';
 import { format, formatDistanceToNow } from 'date-fns';
 import { useState, Fragment } from 'react';
 import { useAttributeListContext } from '@/contexts/attribute-list-context';
+import { LauncherProgressColumn } from '@/components/molecules/session';
+import { FlowProgressColumn } from '@/components/molecules/session';
+import { useEventListContext } from '@/contexts/event-list-context';
+import { ChecklistProgressColumn } from '@/components/molecules/session';
+import { cn } from '@usertour-ui/ui-utils';
+
+const SessionItemContainer = ({
+  children,
+  className,
+}: { children: React.ReactNode; className?: string }) => {
+  return (
+    <div
+      className={cn('flex flex-col w-full px-4 py-6 grow shadow bg-white rounded-lg', className)}
+    >
+      {children}
+    </div>
+  );
+};
 
 interface SessionDetailContentProps {
   environmentId: string;
@@ -23,6 +42,10 @@ export function SessionDetailContent(props: SessionDetailContentProps) {
   const { session } = useQuerySessionDetailQuery(sessionId);
   const [expandedRowId, setExpandedRowId] = useState<string | null>(null);
   const { attributeList } = useAttributeListContext();
+  const { eventList } = useEventListContext();
+  const content = session?.content;
+  const contentType = content?.type;
+  const version = session?.version;
 
   const handleRowClick = (id: string) => {
     setExpandedRowId(expandedRowId === id ? null : id);
@@ -31,6 +54,10 @@ export function SessionDetailContent(props: SessionDetailContentProps) {
   const startEvent = session?.bizEvent?.find(
     (bizEvent) => bizEvent.event?.codeName === BizEvents.FLOW_STARTED,
   );
+
+  if (!eventList || !content || !version) {
+    return <></>;
+  }
 
   return (
     <>
@@ -46,49 +73,63 @@ export function SessionDetailContent(props: SessionDetailContentProps) {
         </div>
       </div>
       <div className="flex flex-col space-y-6 w-full max-w-screen-xl mx-auto p-14 mt-12  ">
-        <div className="flex flex-row px-4 py-6 gap-4 grow shadow bg-white rounded-lg ">
-          <div className=" w-full  grid grid-cols-2 gap-2 gap-x-12">
-            <div className="border-b flex flex-col">
-              <span className="text-sm text-foreground/60">User</span>
-              <Link
-                className="text-primary"
-                to={`/env/${environmentId}/user/${session?.bizUser?.id}`}
-              >
-                {session?.bizUser?.data?.name ?? 'Unnamed user'}
-              </Link>
-            </div>
-            <div className="border-b flex flex-col ">
-              <span className="text-sm text-foreground/60">Flow</span>
-              <Link
-                className=" text-primary"
-                to={`/env/${environmentId}/flows/${session?.content?.id}`}
-              >
-                {session?.content?.name}
-              </Link>
-            </div>
-            <div className="border-b flex flex-col">
-              <span className="text-sm text-foreground/60">Version</span>
-              <Link
-                className="text-primary"
-                to={`/env/${environmentId}/flows/${session?.content?.id}/versions`}
-              >
-                V{session?.version?.sequence}
-              </Link>
-            </div>
-            <div className="border-b flex flex-col">
-              <span className="text-sm text-foreground/60">Started</span>
-              <span>
-                {session?.createdAt && formatDistanceToNow(new Date(session?.createdAt))} ago
-              </span>
-            </div>
-            <div className="border-b flex flex-col">
-              <span className="text-sm text-foreground/60">Start reason</span>
-              <span>{startEvent?.data?.flow_start_reason}</span>
-            </div>
+        <SessionItemContainer className="grid grid-cols-2 gap-2 gap-x-12">
+          <div className="border-b flex flex-col">
+            <span className="text-sm text-foreground/60">User</span>
+            <Link
+              className="text-primary"
+              to={`/env/${environmentId}/user/${session?.bizUser?.id}`}
+            >
+              {session?.bizUser?.data?.name ?? 'Unnamed user'}
+            </Link>
           </div>
-        </div>
+          <div className="border-b flex flex-col ">
+            <span className="text-sm text-foreground/60">Flow</span>
+            <Link
+              className=" text-primary"
+              to={`/env/${environmentId}/flows/${session?.content?.id}`}
+            >
+              {session?.content?.name}
+            </Link>
+          </div>
+          <div className="border-b flex flex-col">
+            <span className="text-sm text-foreground/60">Version</span>
+            <Link
+              className="text-primary"
+              to={`/env/${environmentId}/flows/${session?.content?.id}/versions`}
+            >
+              V{session?.version?.sequence}
+            </Link>
+          </div>
+          <div className="border-b flex flex-col">
+            <span className="text-sm text-foreground/60">Started</span>
+            <span>
+              {session?.createdAt && formatDistanceToNow(new Date(session?.createdAt))} ago
+            </span>
+          </div>
+          <div className="border-b flex flex-col">
+            <span className="text-sm text-foreground/60">Start reason</span>
+            <span>{startEvent?.data?.flow_start_reason}</span>
+          </div>
+        </SessionItemContainer>
+        <SessionItemContainer>
+          <div className="mb-2 flex flex-row items-center font-bold	">
+            <PieChartIcon width={18} height={18} className="mr-2" />
+            Progress
+          </div>
+          {contentType === ContentDataType.CHECKLIST && (
+            <ChecklistProgressColumn original={session} eventList={eventList} version={version} />
+          )}
 
-        <div className="flex flex-col w-full px-4 py-6 grow shadow bg-white rounded-lg">
+          {contentType === ContentDataType.FLOW && (
+            <FlowProgressColumn original={session} eventList={eventList} />
+          )}
+          {contentType === ContentDataType.LAUNCHER && (
+            <LauncherProgressColumn original={session} eventList={eventList} />
+          )}
+        </SessionItemContainer>
+
+        <SessionItemContainer>
           <div className="mb-2 flex flex-row items-center font-bold	">
             <ActivityLogIcon width={18} height={18} className="mr-2" />
             Activity feed
@@ -144,7 +185,7 @@ export function SessionDetailContent(props: SessionDetailContentProps) {
               </TableBody>
             </Table>
           </div>
-        </div>
+        </SessionItemContainer>
       </div>
     </>
   );
